\name{minb_SimuData}
\alias{minb_SimuData}
\title{An Example of Simulated Data for minb }
\docType{data}
\description{
The dataset minb_SimuData contains n = 2000 samples simulated from multiple inflated negative binomial model with p = 15 predictors while the first 10 predictors are informative. The responses are inflated at multiple integers 0, 1, 3, 5, and 10.
}
\usage{minb_SimuData}
\format{A data list containing 2000 samples}
